pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraudit"
version = "0.1.0"
description = "Texas government spending fraud detection and analysis system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Fraudit Contributors"}
]
keywords = [
    "fraud-detection",
    "government-spending",
    "texas",
    "audit",
    "transparency",
    "procurement",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]

dependencies = [
    # CLI
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Database
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.12",

    # Data ingestion
    "sodapy>=2.2",
    "requests>=2.31",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "beautifulsoup4>=4.12",

    # Analysis
    "rapidfuzz>=3.0",
    "networkx>=3.0",
    "numpy>=1.24",

    # Scheduling
    "apscheduler>=3.10",

    # Utilities
    "tqdm>=4.66",
    "tabulate>=0.9",

    # Terminal UI
    "textual>=0.47",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/fraudit"
Documentation = "https://github.com/yourusername/fraudit#readme"
Repository = "https://github.com/yourusername/fraudit.git"
Issues = "https://github.com/yourusername/fraudit/issues"

[project.scripts]
fraudit = "fraudit.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["fraudit*"]

[tool.setuptools.package-data]
fraudit = [
    "pia/templates/**/*",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
